backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xentraldesk-backend"
version = "0.1.0"
description = "XentralDesk — AI-powered customer support backend"
readme = "README.md" 
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
    "alembic>=1.14.0",
    # Settings & validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.1.0",
    # Auth & security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Redis
    "redis[hiredis]>=5.2.0",
    # HTTP client
    "httpx>=0.28.0",
    # AI
    "google-genai>=1.0.0",
    "openai>=1.0.0",
    # Google OAuth & Gmail API
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    # Payments
    "stripe>=11.0.0",
    # Utils
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.1",
    "discord.py>=2.4.0",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "arq>=0.26.0",
    # Chunking & embeddings
    "tiktoken>=0.5.0",
    "tenacity>=8.2.0",
    # Web scraping
    "beautifulsoup4>=4.12.0",
    "playwright>=1.47.0",
    "trafilatura>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]


[tool.hatch.build.targets.wheel]
packages = ["app"]